requirements.txt
streamlit
ultralytics
pillow
numpy
matplotlib
opencv-python-headless
